make staticcheck happy
